# lib/api-zod/src/generated/types/createStudySetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyForge AI API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateStudySetBodyDifficulty } from "./createStudySetBodyDifficulty";

export interface CreateStudySetBody {
  title: string;
  subject: string;
  /** @nullable */
  course?: string | null;
  difficulty: CreateStudySetBodyDifficulty;
  /** @nullable */
  learningGoal?: string | null;
  /** @nullable */
  examDate?: string | null;
  notes: string;
}